+++ projects/igniteui-angular/src/lib/combo/combo.pipes.ts
@@ -1,14 +1,12 @@
 import { Inject, Pipe, PipeTransform } from '@angular/core';
 import { cloneArray } from '../core/utils';
 import { DataUtil } from '../data-operations/data-util';
-import { IGX_COMBO_COMPONENT, IgxComboBase } from './combo.common';
 import { DefaultSortingStrategy, SortingDirection } from '../data-operations/sorting-strategy';
+import { IgxComboBase, IGX_COMBO_COMPONENT } from './combo.common';
 import { IComboFilteringOptions } from './combo.component';
 
 /** @hidden */
-@Pipe({
-    name: 'comboClean'
-})
+@Pipe({ name: 'comboClean' })
 export class IgxComboCleanPipe implements PipeTransform {
     public transform(collection: any[]) {
         return collection.filter(e => !!e);
@@ -16,27 +14,25 @@ export class IgxComboCleanPipe implements PipeTransform {
 }
 
 /** @hidden */
-@Pipe({
-    name: 'comboFiltering'
-})
+@Pipe({ name: 'comboFiltering' })
 export class IgxComboFilteringPipe implements PipeTransform {
-    public transform(collection: any[], searchValue: any, displayKey: any,
-        filteringOptions: IComboFilteringOptions, shouldFilter = false) {
+    private displayKey: any;
+
+    public transform (
+        collection: any[],
+        searchValue: any,
+        displayKey: any,
+        filteringOptions: IComboFilteringOptions,
+        filterable: boolean,
+        filterFunction: (collection: any[], searchValue: any, caseSensitive: boolean) => any[] = defaultFilterFunction) {
         if (!collection) {
             return [];
         }
-        if (!searchValue || !shouldFilter) {
+        if (!filterable) {
             return collection;
-        } else {
-            const searchTerm = filteringOptions.caseSensitive ? searchValue.trim() : searchValue.toLowerCase().trim();
-            if (displayKey != null) {
-                return collection.filter(e => filteringOptions.caseSensitive ? e[displayKey]?.includes(searchTerm) :
-                    e[displayKey]?.toString().toLowerCase().includes(searchTerm));
-            } else {
-                return collection.filter(e => filteringOptions.caseSensitive ? e.includes(searchTerm) :
-                    e.toString().toLowerCase().includes(searchTerm));
-            }
         }
+        this.displayKey = displayKey;
+        return filterFunction.call(this, collection, searchValue, filteringOptions.caseSensitive);
     }
 }
 
@@ -78,3 +74,20 @@ export class IgxComboGroupingPipe implements PipeTransform {
         return data;
     }
 }
+
+function defaultFilterFunction (collection: any[], searchValue: any, matchCase: boolean): any[] {
+    if (!searchValue) {
+        return collection;
+    }
+    const searchTerm = matchCase ? searchValue.trim() : searchValue.toLowerCase().trim();
+    if (this.displayKey != null) {
+        return collection.filter(e => matchCase ?
+            e[this.displayKey]?.includes(searchTerm) :
+            e[this.displayKey]?.toString().toLowerCase().includes(searchTerm));
+    } else {
+        return collection.filter(e => matchCase ?
+            e.includes(searchTerm) :
+            e.toString().toLowerCase().includes(searchTerm));
+    }
+}
+