+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 All notable changes for each version of this project will be documented in this file.
 
+## 14.0.6
+
+### New Features
+- `IgxCombo` and  `IgxSimpleComboComponent`
+    - `filterFunction` input is added. The new property allows changing of the way filtering is done in the combos. By default filtering is made over the values in combo's data when it is a collection of primitive values, or over the values as defined in `displayKey` of the combo. If custom filtering function is provided filtering will be done as specified in the provided function.
+
 ## 14.0.0
 
 - `IgxDatePicker` and `IgxDateRangePicker` now expose a `weekStart` input property like the `IgxCalendar`
-
-### New Features
 
 ### General
 - Updating dependency to Angular 14